src/preprocess.py
import numpy as np
from PIL import Image

def preprocess_image(image: Image.Image):
    image = image.resize((224, 224))
    image = np.array(image, dtype=np.float32) / 255.0
    image = np.expand_dims(image, axis=0)
    return image
