requirements.txt
tensorflow==2.13.0
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
mlflow==2.10.2
fastapi==0.103.2
uvicorn==0.23.2
pillow==10.2.0
typing-extensions==4.5.0
python-multipart==0.0.9

